+++ CHANGELOG.md
@@ -6,4 +6,11 @@
 
 ## 221030
 
-* First alpha release. Mostly very stupid bug fixed.
+* First alpha release. Mostly bug fixing.
+
+# 221101
+
+* Better retry logic for commands
+* Rain Delay with dedicated actions (x_opensprinkler_raindelay.set)
+* parsing logic for mqtt integration
+* Refactoring